fix: Replay network detail unrecognized selector
The @objc setRequest/setResponse/serialize methods on
SentryReplayNetworkDetails lived in a Swift extension, which compiles to
ObjC category metadata. The static linker strips categories without
-ObjC/-force_load, so in statically linked distributions (e.g. React
Native) those selectors were missing at runtime. The first setRequest
call during Session Replay network-detail capture then crashed with
NSInvalidArgumentException: unrecognized selector sent to instance.

init(method:) survived because it lives in the main class body and rides
in the class's own method list, which explains why the object was created
but died on the first capture call.

Move the @objc methods back into the main class body (the pre-9.25
layout) so they are part of the primary method list and never stripped.

Fixes getsentry/sentry-react-native#6609
